registration/pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "doccollate-registration"
version = "0.1.0"
description = "Standalone registration-form generator"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2,<3",
  "python-docx",
  "python-dotenv",
  "pyyaml",
  "tomli; python_version < '3.11'",
]

[project.scripts]
registration-form = "registration_form.cli:main"

[tool.hatch.build]
include = [
  "src/registration_form/resources/*",
]

[tool.hatch.build.targets.wheel]
packages = ["src/registration_form"]

[tool.doccollate.templates]
registration = "src/registration_form/resources/test_registration_form.docx"

[tool.doccollate.doccollate]
contact_info = "src/registration_form/resources/soft_copyright.yaml"
preset_choice = ""
